requirements.txt
asgiref
attrs
Django>=3.2,<3.3
django-cors-headers
django-filter
djangorestframework
drf-spectacular
flake8
inflection
jsonschema
jsonschema-specifications
mccabe
pillow
pycodestyle
PyMySQL
python-dotenv
pytz
PyYAML
referencing
rpds-py
sqlparse
typing_extensions
tzdata
uritemplate
cryptography
djangorestframework_simplejwt
python-dateutil
APScheduler
requests